fix session problem with Lists

+++ action_delete_item.php
@@ -11,12 +11,10 @@
     return;
   }
 
-  if(isset($_SESSION['username']) && $_SESSION['username'] != ''){
-    if(ListBelongsUser($dbh,$_SESSION['username'],$item['idList'])) {
+  if(isset($_SESSION['username']) && $_SESSION['username'] != '' && ListBelongsUser($dbh,$_SESSION['username'],$item['idList'])){
       deleteItem($dbh,$_GET["item"]);
       updateModified($dbh,$item['idList']);
       echo 0;
-    }
   }
   else {
     echo -1;
