+++ action_delete_list.php
@@ -11,12 +11,10 @@
     return;
   }
 
-  if(isset($_SESSION['username']) && $_SESSION['username'] != ''){
-    if(ListBelongsUser($dbh,$_SESSION['username'],$idList)) {
+  if(isset($_SESSION['username']) && $_SESSION['username'] != '' && ListBelongsUser($dbh,$_SESSION['username'],$idList)){
       deleteList($dbh,$idList);
       echo 0;
     }
-  }
   else {
     echo -1;
   }
